feat: Implement shop management API with modular routing
- Introduced a new shop router module, organizing routes into `admin_only` and `all_users` components.
- The `admin_only` module provides CRUD operations for shops, enforcing admin access.
- The `all_users` module offers a paginated list of shops for any authenticated user.
- Improved code organization and maintainability by separating concerns within the shop management functionality.
